+++ plugins/optimizer/coordinator/routes.py
@@ -28,6 +28,7 @@ from flask.views import MethodView
 from marshmallow import EXCLUDE
 
 from plugins.optimizer.coordinator.tasks import echo_results, get_features_and_target
+from plugins.optimizer.interaction_utils.tasks import invoke_task
 from qhana_plugin_runner.api.plugin_schemas import (
     DataMetadata,
     EntryPoint,
@@ -35,10 +36,8 @@ from qhana_plugin_runner.api.plugin_schemas import (
     PluginMetadataSchema,
     PluginType,
 )
-from qhana_plugin_runner.api.tasks_api import TaskData, TaskStatusSchema
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
-from qhana_plugin_runner.storage import STORE
-from qhana_plugin_runner.tasks import invoke_task, save_task_error, save_task_result
+from qhana_plugin_runner.tasks import save_task_error, save_task_result
 
 from . import OPTIMIZER_BLP, Optimizer
 from .schemas import OptimizerSetupTaskInputData, OptimizerSetupTaskInputSchema
